fix: 언마운트 시 타이머 cleanup 추가 및 코드 분리
